feat(ui): create multisign error page
