+++ src/ui/pages/NotificationDetails/components/MultiSigRequest/MultiSigRequest.scss
