+++ src/ui/pages/NotificationDetails/components/MultiSigRequest/index.ts
@@ -1,0 +1,1 @@
+export * from "./MultiSigRequest";
